requirements.txt
torchvision
tensorflow-gpu==1.14
torchcontrib
pytest
matplotlib
scipy==1.2.0
sklearn
protobuf==3.11.2